pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "plutosdr-driver"
version = "0.1.0"
description = "Python driver helpers for PlutoSDR"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pyadi-iio",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]